requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
facenet-pytorch>=2.5.2

# Web application
Flask>=2.0.0

# Image processing
Pillow>=8.0.0
opencv-python>=4.5.0

# Data science utilities
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.50.0

# For path and compatibility
pathlib

# Optional: For Google Fonts in HTML preview (no pip install needed)
# For local development (Jupyter notebooks)
ipykernel>=6.0.0
notebook>=6.0.0
